requirements.txt
# ─── Core AI/ML ───────────────────────────────────────────────────────
sentence-transformers>=2.7.0      # Open-source embeddings (all-MiniLM-L6-v2)
torch>=2.0.0                      # Required by sentence-transformers (CPU ok)
transformers>=4.40.0              # Optional: LLM answer generation
numpy>=1.24.0                     # Vector math utilities

# ─── Endee Vector Database SDK ────────────────────────────────────────
endee>=0.1.0                      # Official Python SDK

# ─── NLP Utilities ────────────────────────────────────────────────────
nltk>=3.8.1                       # Tokenization helpers (optional)

# ─── Testing ──────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=5.0.0

# ─── Dev / Quality ────────────────────────────────────────────────────
black>=24.0.0
ruff>=0.4.0
groq
google-genai